--- augments/modules/wikipedia/__init__.py
@@ -1,1 +1,0 @@
-from .text_crawler import *
